fix:二叉树学习思路篇--树展开为单链表&有序数组恢复成二叉搜索树

+++ sun-summarize-leetcode/src/main/java/record/leetcode/editor/cn/PathSumIii.java
@@ -68,7 +68,7 @@ public class PathSumIii {
 	 */
 	class Solution {
 		// 路径和
-		private int pathSum = 0;
+		private long pathSum = 0;
 		// 路径和结果集
 		private int result = 0;
 		// 目标值
